+++ History.md
@@ -2,6 +2,7 @@
 ==================
   * Re-package forked Rhino for cleaner packaging (https://github.com/tntim96/JSCover/issues/130)
   * Add shutdown hook
+  * Upgrade to Jasmine 2, HtmlUnit 2.15 (https://github.com/HtmlUnit/htmlunit-rhino-fork/pull/2)
   * Internal: Minor modifications to assist with Maven plugin
 
 1.0.9 / 2014-05-23